feat: add provideAuth for standalone applications
